test(logic): split game_setup_creation part2 for file-size gate (Refs #2288)
Move OW/NW assignment balance tests into part2_assignment_test.dart so
both files stay under the 400-line logic test gate. No behavior change.
